Step 3: Extract hostname false-positive filters to hostname_filters.py
Moves all FP filter functions and constants out of urls.py into a dedicated
hostname_filters.py module. The function is renamed to is_false_positive_hostname
(removing the leading underscore since it is now a module-level API).
